+++ test/e2e/lib.ts
@@ -105,7 +105,7 @@ function dumpTest(
         }
 
         // remove the file
-        fs.unlinkSync(filename);
+        if(fs.existsSync(filename)) fs.unlinkSync(filename);
 
         // ASSERT
         const memoryLines = [];
